Extract WorkQueue to its own module
