+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod template_transformers;
 mod tree;
+mod work_queue;
 
 pub mod links;
 pub mod wikitext_parser;
